requirements.txt
crispy-bootstrap5
django>=4.2,<5.0
django-crispy-forms
structlog